--- a/src/LightGraphs.jl
+++ b/src/LightGraphs.jl
@@ -48,7 +48,7 @@ crosspath,
 # graph visit
 SimpleGraphVisitor, TrivialGraphVisitor, LogGraphVisitor,
 discover_vertex!, open_vertex!, close_vertex!,
-examine_neighbor!, visited_vertices, traverse_graph, traverse_graph_withlog,
+examine_neighbor!, visited_vertices, traverse_graph!, traverse_graph_withlog,
 
 # bfs
 BreadthFirst, gdistances, gdistances!, bfs_tree, is_bipartite, bipartite_map,