--- a/src/connectivity.jl
+++ b/src/connectivity.jl
@@ -259,7 +259,7 @@ end
 
 Returns a vector of the vertices in `g` at distance less or equal to `d`
 from `v`. If `g` is a `DiGraph` the `dir` optional argument specifies the edge direction
-with respect to `v` (i.e. `:in` or `:out`) to be considered.
+the edge direction with respect to `v` (i.e. `:in` or `:out`) to be considered.
 """
 function neighborhood(g::SimpleGraph, v::Int, d::Int; dir=:out)
     @assert d >= 0 "Distance has to be greater then zero."
@@ -269,3 +269,14 @@ function neighborhood(g::SimpleGraph, v::Int, d::Int; dir=:out)
         vertexcolormap=Dict{Int,Int}(), dir=dir)
     return visitor.neigs
 end
+
+
+"""
+    egonet(g, v::Int, d::Int; dir=:out)
+
+Returns the subgraph of `g` induced by the neighbors of `v` up to distance
+`d`. If `g` is a `DiGraph` the `dir` optional argument specifies
+the edge direction the edge direction with respect to `v` (i.e. `:in` or `:out`)
+to be considered. This is equivalent to `induced_subgraph(g, neighborhood(g, v, d, dir=dir)).`
+"""
+egonet(g::SimpleGraph, v::Int, d::Int; dir=:out) = induced_subgraph(g, neighborhood(g, v, d, dir=dir))