--- a/src/connectivity.jl
+++ b/src/connectivity.jl
@@ -24,7 +24,7 @@ function connected_components!(label::Vector{Int}, g::SimpleGraph)
     # passed to components(a)
     nvg = nv(g)
     visitor = LightGraphs.ComponentVisitorVector(label, 0)
-    colormap = fill(-1,nvg)
+    colormap = fill(0, nvg)
     queue = Vector{Int}()
     sizehint!(queue, nvg)
     for v in 1:nvg
@@ -133,7 +133,7 @@ function discover_vertex!(vis::TarjanVisitor, v)
 end
 
 function examine_neighbor!(vis::TarjanVisitor, v, w, w_color::Int, e_color::Int)
-    if w_color >= 0 # >=0 means added seen
+    if w_color != 0 # != 0 means seen
         while vis.index[w] > 0 && vis.index[w] < vis.lowlink[end]
             pop!(vis.lowlink)
         end
@@ -246,8 +246,8 @@ end
 NeighborhoodVisitor(d::Int) = NeighborhoodVisitor(d, Vector{Int}())
 
 function examine_neighbor!(visitor::NeighborhoodVisitor, u::Int, v::Int, ucolor::Int, vcolor::Int, ecolor::Int)
-    ucolor >= visitor.d && return false
-    if vcolor < 0
+    -ucolor > visitor.d && return false # color is negative for not-closed vertices
+    if vcolor == 0
         push!(visitor.neigs, v)
     end
     return true