as a follow up of the discussion in #329 

the crucial line here is the one that uses `get` to treat on the same footing arrays and dictionaries

~~the "add neighborhood" commit got in by mistake, I'll remove it later~~
this also adds the neighborhood and egonet functions, based on breadth first search.  

this also fixes #330