@CarloLucibello 2016-04-12
I made Dict{Int,Int}() the default for vertexcolormap, since some experiments with large erdos_renyi graph and `is_bipartite` (which also relies on `connected_components`) showed up to 4x performance improvements in some cases and never a regression.