@jpfairbanks 2016-04-12
I think we should have three methods on this function: one with no type constraints, another with an AbstractArray{T}, and the third with Dict{Int,T}. The current code can go in the unconstrained method and the other two methods would provide the "types as documentation" that @sbromberger wants.