@sbromberger 2016-04-12
@jpfairbanks My initial reaction is that this is probably the worst of all possible solutions.

I'd MAYBE be ok with `Associative` and `AbstractArray`, as long as one converted to the other (so we don't have specialized code).

I don't see the use case, though. Aside from some minor additional convenience, why are `Int`-based colormaps not workable? What's next, changing vertex types?