@jpfairbanks 2016-04-12
Do you dislike the type parameterization `T` or the three definitions?