@sbromberger 2016-04-12
I'm honestly not a fan of either. What's wrong with integer colors? We use ints for coloring all over the place: partitions, biconnected components, .... this change will introduce an inconsistency AND make things more complex without an identified use case.