@jpfairbanks 2016-04-13
I will leave the contained type for now and say it can stay as ints.

Making the color map able to support both Dict and Vector is important for exactly the use case carlo has found. If you want to do a BFS but only go a constant number of hops, then you shouldn't allocate o(nv) memory. That is why we need to support some sparse vector for the colormap. Carlo has some experiments that show Dict works well enough.