@sbromberger 2016-04-13
But here's the thing: if you don't want to allocate `nv` Ints (and, let's face it, for 1 million nodes that's 64 MB) because you're concerned about memory consumption, I don't see how a Dict is going to help you, since you're at the same allocation if you traverse more than `nv / 4` nodes. This becomes even worse when you use data types that are larger than 64 bits.

I'm missing something. What, specifically is the use case with the limitation?

All else being equal, we want to limit memory allocation. However, in this case, we're faced with a tradeoff: get more efficient memory allocs in limited, specific defined cases, but (on the other hand) increase the code complexity AND create a situation where we get HORRIBLE memory allocs in some other defined cases. I'm not sure this is a good tradeoff.