@CarloLucibello 2016-04-13
the problem to me is not memory consumption, it is performance: the reason I started implementing neighborhood is that I need it inside an algorithm that has to call it many many times. Now, each call would imply allocating and deallocating O(nv) memory, and this would take most of the computational time. 

Unless we want to optionally pass to neighborhood  an initialized  colormap, which would be really a questionable design choice, the only solution here is to allow both dicts and vectors for colormap. 

I don't really see any problem in this: 
what we gain? more flexibility, reduced memory consumption and improved or on par performance on all case that I tested (not many, but still somenthing)
what we lose? code complexity? not true, I just had to change two line of code to generalize graph_traversal. Horrible memory allocation in some cases? I experienced quite the opposite:
```julia
####
julia> g=erdos_renyi(100_000,100_000)
{100000, 100000} undirected graph

# BEFORE
julia> @time is_bipartite(g)
  0.045502 seconds (84.06 k allocations: 10.419 MB, 21.31% gc time)
false

julia> @time is_bipartite(g)
  0.034474 seconds (84.06 k allocations: 10.419 MB)
false

#AFTER
julia> @time is_bipartite(g)
  0.037780 seconds (84.96 k allocations: 9.797 MB)
false

julia> @time is_bipartite(g)
  0.036506 seconds (84.96 k allocations: 9.797 MB)
false
#####
julia> g=erdos_renyi(1_000_000,10_000_000)
{1000000, 10000000} undirected graph

#BEFORE
julia> @time is_bipartite(g)
  1.038482 seconds (74 allocations: 40.599 MB, 0.41% gc time)
false

julia> @time is_bipartite(g)
  1.041845 seconds (74 allocations: 40.599 MB, 0.31% gc time)
false

#AFTER
julia> @time is_bipartite(g)
  1.086443 seconds (119 allocations: 33.325 MB, 0.21% gc time)
false

julia> @time is_bipartite(g)
  1.074744 seconds (119 allocations: 33.325 MB, 0.13% gc time)
false

# case in which it is actually bipartite
julia> g=erdos_renyi(100_000,10_000)
{100000, 10000} undirected graph

#BEFORE
julia> @time is_bipartite(g)
  0.232985 seconds (481.32 k allocations: 1.243 GB, 18.40% gc time)
true

julia> @time is_bipartite(g)
  0.278150 seconds (481.32 k allocations: 1.243 GB, 31.56% gc time)

#AFTER
julia> @time is_bipartite(g)
  0.099805 seconds (484.05 k allocations: 167.945 MB, 34.27% gc time)
true

julia> @time is_bipartite(g)
  0.080207 seconds (484.05 k allocations: 167.945 MB, 20.56% gc time)
true
```
This results are not clear to me, they should be investigated, anyway in no case there is a regression and in one case there is a huge improvements using dicts.