@sbromberger 2016-04-13
@CarloLucibello your numbers are compelling, and my tests bear out what you're saying. Thanks for sticking with this.

Please see more comments in code. I think we can do this so that it remains consistent with LG standards (which should probably be codified at some point).

@jpfairbanks - I think I've been able to do this without type parameterization, of which I'm not a huge fan. Take a look?